+++ src/build.js
@@ -1,6 +1,5 @@
 const fs = require('fs');
 const process = require('process');
-const commander = require('commander').program;
 const log = require('@vladmandic/pilogger');
 const compile = require('./compile.js');
 const watch = require('./watch.js');
@@ -12,6 +11,7 @@ const helpers = require('./helpers');
 const defaults = require('../build.json');
 const clean = require('./clean.js');
 const changelog = require('./changelog.js');
+const cli = require('./cli.js');
 const app = require('../package.json');
 
 process.on('SIGINT', () => {
@@ -43,13 +43,6 @@ const updateConfig = (config, options) => {
   }
   if (Object.keys(options).length) local = helpers.merge(local, options);
   return local;
-};
-
-const printInfo = (type, application, environment, toolchain) => {
-  log.info('Application:', application);
-  log.info('Environment:', { profile: type, ...environment });
-  log.info('Toolchain:', toolchain);
-  // log.data('Configuration:', config);
 };
 
 class Build {
@@ -114,7 +107,7 @@ class Build {
   /**
    * Initializes Build class with all parsed configurations
    *
-   * @param options  Optional configuration options overrides
+   * @param options  `<object>` Optional configuration options overrides
    */
   constructor(options = {}) {
     this.config = updateConfig(helpers.merge(defaults), options);
@@ -125,30 +118,34 @@ class Build {
     this.toolchain = { build: app.version, esbuild: compile.version, typescript: typings.version, typedoc: typedoc.version, eslint: lint.version };
     this.environment = { config: 'build.json', tsconfig, eslintrc, git };
     this.application = { name: this.package.name, version: this.package.version };
+    log.ringLength = 1000; // increase log ring buffer
   }
 
   /**
    * Runs build pipeline for development profile
    *
-   * @param options  optional configuration options overrides
+   * @param options  `<object>` Optional configuration options overrides
+   * @returns  `array<object>` Containing all messages
    */
   async development(options = {}) {
     if (Object.keys(options).length) this.config = updateConfig(this.config, options);
-    printInfo('development', this.application, this.environment, this.toolchain);
+    helpers.info('development', this.application, this.environment, this.toolchain);
     if (this.config.debug) log.data('Configuration:', this.config);
     if (this.config.serve.enabled) await serve.start(this.config.serve);
     if (this.config.watch.enabled) await watch.start(this.config);
     if (this.config.build.enabled) await compile.build(this.config, { type: 'development' });
+    return helpers.results();
   }
 
   /**
    * Runs build pipeline for production profile
    *
-   * @param options  optional configuration options overrides
+   * @param options  `<object>` Optional configuration options overrides
+   * @returns  `array<object>` Containing all messages
    */
   async production(options = {}) {
     if (Object.keys(options).length) this.config = updateConfig(this.config, options);
-    printInfo('production', this.application, this.environment, this.toolchain);
+    helpers.info('production', this.application, this.environment, this.toolchain);
     if (this.config.debug) log.data('Configuration:', this.config);
     if (this.config.clean.enabled) await clean.start(this.config.clean);
     if (this.config.build.enabled) await compile.build(this.config, { type: 'production' });
@@ -156,82 +153,27 @@ class Build {
     // if (config.typings.enabled) await typings.run(config.typings, entry.input); // triggered from compile.build
     // if (config.typedoc.enabled) await typedoc.run(config.typedoc, entry.input); // triggered from compile.build
     if (this.config.changelog.enabled && this.config.changelog.output && this.config.changelog.output !== '' && this.environment.git) await changelog.update(this.config, this.package); // generate changelog
-    log.info('Profile production done');
+    return helpers.results();
   }
 
   /**
    * Runs build pipeline for specified profile
    *
-   * @param profile  profile type: <"production" | "development">
-   * @param options  optional configuration options overrides
+   * @param profile  `string` Profile type: "production" | "development"
+   * @param options  `object` Optional configuration options overrides
+   * @returns  `array<object>` Containing all messages
    */
-  build(profile = '', options = {}) {
-    if (profile === 'production') this.production(options);
-    else if (profile === 'development') this.development(options);
+  async build(profile = '', options = {}) {
+    let res = [];
+    if (profile === 'production') res = await this.production(options);
+    else if (profile === 'development') res = await this.development(options);
     else log.error('Build:', 'unknonwn profile');
-  }
-
-  /**
-   * Runs build in cli mode
-   *
-   * Usage: build [options] [command]
-   *
-   * **Options:**
-   * - `-c`, `--config` `<file>`  specify alternative config file
-   * - `-d`, `--debug`            enable debug output
-   * - `-h`, `--help`             display help for command
-   *
-   * **Commands:**
-   * - `development`          start development ci
-   * - `production`           start production build
-   * - `config`               show active configuration and exit
-   * - `help [command]`       display help for command
-   */
-  cli() {
-    if (this.config.log.enabled && this.config.log.output && this.config.log.output !== '') log.logFile(this.config.log.file);
-
-    log.header();
-    if (this.environment.tsconfig) this.config.build.global.tsconfig = 'tsconfig.json';
-    commander.option('-c, --config <file>', 'specify config file');
-    commander.option('-d, --debug', 'enable debug output');
-    commander.command('development').description('start development ci').action(() => this.params.command = 'development');
-    commander.command('production').description('start production build').action(() => this.params.command = 'production');
-    commander.command('config').description('show active configuration and exit').action(() => this.params.command = 'config');
-    commander.parse(process.argv);
-    this.params = { ...this.params, ...commander.opts() };
-    if (this.params.debug) {
-      log.info('Enabling debug output');
-      this.config.debug = true;
-    }
-    if (this.params.config && this.params.config !== '') {
-      if (fs.existsSync(this.params.config)) {
-        const data = fs.readFileSync(this.params.config);
-        try {
-          this.config = helpers.merge(this.config, JSON.parse(data.toString()));
-          log.info('Parsed config file:', this.params.config, this.config);
-        } catch {
-          log.error('Error parsing config file:', this.params.config);
-        }
-      } else {
-        log.error('Config file does not exist:', this.params.config);
-      }
-    }
-    switch (this.params.command) {
-      case 'development': this.development(); break;
-      case 'production': this.production(); break;
-      case 'config': {
-        printInfo('production', this.application, this.environment, this.toolchain);
-        log.data('Configuration:', this.config);
-        break;
-      }
-      default:
-    }
+    return res;
   }
 }
 
 exports.Build = Build;
 
 if (require.main === module) {
-  const build = new Build();
-  build.cli();
+  cli.run();
 }