+++ src/log.js
@@ -1,0 +1,3 @@
+const log = require('@vladmandic/pilogger');
+
+for (const key of Object.keys(log)) exports[key] = log[key];