+++ types/src/helpers.d.ts
@@ -1,1 +1,3 @@
 export function merge(...objects: any[]): any;
+export function info(type: any, application: any, environment: any, toolchain: any): void;
+export function results(): any[];